backend/video/requirements.txt
google-cloud-aiplatform[all]==1.69.0
Flask==3.0.3
gunicorn==22.0.0
Flask-Cors==5.0.0
requests==2.32.3
IPython==8.31.0